+++ packages/common/tests/setup/connection.ts
@@ -1,3 +1,4 @@
+import logger from '../../src/util/logger'
 const Connection = require('../../src/services/Connection')
 jest.mock('../../src/services/Connection')
 
@@ -7,3 +8,6 @@ jest.mock('uuid', () => {
     v4: jest.fn(() => 'mocked-uuid')
   };
 });
+
+
+logger.disableAll()